pyproject.toml
[project]
name = "job-application-tracker"
version = "1.0.0"
description = "AI-powered job application tracker with automated scraping and matching"
requires-python = ">=3.12"
dependencies = [
    "streamlit>=1.52.2",
    "pandas>=2.3.3",
    "python-jobspy>=1.1.82",
    "langgraph>=1.0.5",
    "openai>=2.13.0",
    "plotly>=6.5.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "simhash>=2.1.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
